+++ go/path_srv/internal/psconfig/psconfig.go
@@ -18,7 +18,7 @@ package psconfig
 import (
 	"time"
 
-	"github.com/scionproto/scion/go/lib/env"
+	"github.com/scionproto/scion/go/lib/util"
 )
 
 var (
@@ -32,7 +32,7 @@ type Config struct {
 	PathDB  string
 	// QueryInterval specifies after how much time segments
 	// for a destination should be refetched.
-	QueryInterval env.Duration
+	QueryInterval util.DurWrap
 }
 
 func (c *Config) InitDefaults() {
